Improve setup

+++ setup.sh
@@ -2,22 +2,47 @@
 
 set -e
 
-/bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
+if [[ -d ${HOME}/.config ]]; then
+	rm -rf ${HOME}/.config.backup
+	mv ${HOME}/.config ${HOME}/.config.backup
+	mkdir ${HOME}/.config
+fi
 
+if [[ -f ${HOME}/.zshrc ]]; then 
+	mv ${HOME}/.zshrc ${HOME}/.zshrc.bak
+fi
 
-mv -r ${HOME}/.config ${HOME}/.config.backup
-mv ${HOME}/.zshrc ${HOME}/.zshrc.bak
-mkdir .config
+if [[ $(uname) == "Darwin" ]]; then
+	/bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
+	brew install iterm2 python3 python3-pip python3-venv tmux neovim fzf ripgrep fd llvm jq git-lfs exa ncdu bottom cmake unzip thefuck
+elif command -v apt > /dev/null; then
+    sudo add-apt-repository ppa:neovim-ppa/unstable
+	sudo apt update
+	sudo apt install -y zsh python3 python3-venv python3-pip unzip cmake tmux neovim fzf ripgrep fd-find llvm jq git-lfs exa ncdu thefuck
+	curl -LO https://github.com/ClementTsang/bottom/releases/download/0.9.6/bottom_0.9.6_amd64.deb
+	sudo dpkg -i bottom_0.9.6_amd64.deb
+	chsh -s $(which zsh)
+	if [[ ! -d ${HOME}/.local/bin ]]; then
+		mkdir -p ${HOME}/.local/bin
+	fi
+	if [[ -L ${HOME}/.local/bin/fd ]]; then
+		rm ${HOME}/.local/bin/fd
+	fi
+	ln -s /usr/bin/fdfind ${HOME}/.local/bin/fd
+fi
 
-sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
-
-brew install iterm2 tmux neovim fzf ripgrep fd llvm jq git-lfs exa ncdu
+rm -rf ${HOME}/.oh-my-zsh
+bash -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)" "" --unattended
 
 git clone --depth=1 https://github.com/romkatv/powerlevel10k.git ${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}/themes/powerlevel10k
-git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting
-git clone https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
-git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
-git clone --depth 1 https://github.com/wbthomason/packer.nvim ~/.local/share/nvim/site/pack/packer/start/packer.nvim
+git clone --depth=1 https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-${HOME}/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting
+git clone --depth=1 https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-${HOME}/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
+git clone --depth=1 https://github.com/ptavares/zsh-exa.git ${ZSH_CUSTOM:-${HOME}/.oh-my-zsh/custom}/plugins/zsh-exa
+rm -rf ${HOME}/.tmux
+git clone https://github.com/tmux-plugins/tpm ${HOME}/.tmux/plugins/tpm
+
+rm -rf  ${HOME}/.local/share/nvim/site/pack/packer/start/packer.nvim
+git clone --depth=1 https://github.com/wbthomason/packer.nvim ${HOME}/.local/share/nvim/site/pack/packer/start/packer.nvim
 
 echo 'source ${HOME}/.config/shell/aliases.zsh'>> ${HOME}/.zshrc
 echo 'source ${HOME}/.config/shell/exports.zsh'>> ${HOME}/.zshrc
@@ -26,12 +51,11 @@ echo 'source ${HOME}/.config/shell/p10k.zsh'>> ${HOME}/.zshrc
 echo 'source ${HOME}/.config/fzf/completion.sh'>> ${HOME}/.zshrc
 echo 'source ${HOME}/.config/fzf/keybindings.sh'>> ${HOME}/.zshrc
 
-cp -r ./fzf ${HOME}/.config/fzf
-cp -r ./shell ${HOME}/.config/shell
-cp -r ./nvim ${HOME}/.config/nvim
-cp -r ./tmux ${HOME}/.config/tmux
+cp -r ./fzf ${HOME}/.config
+cp -r ./shell ${HOME}/.config
+cp -r ./nvim ${HOME}/.config
+cp -r ./tmux ${HOME}/.config
 
-source "${HOME}"/.zshrc
-
-
+#zsh
+#source "${HOME}"/.zshrc
